requirements.txt
alabaster==0.7.12
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.3
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
django-redis==4.12.1
djangorestframework==3.11.0
docutils==0.16
dodgy==0.2.1
drf-yasg==1.17.1
flake8==3.8.3
flake8-polyfill==1.0.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.7.0
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
pep8-naming==0.10.0
prospector==1.3.0
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pylint-celery==0.3
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
requirements-detector==0.7
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
setoptconf==0.2.0
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
